+++ src/app/config/api-config.ts
@@ -26,6 +26,7 @@ export interface ApiConfig {
   internalSearchRoute: string;
   externalSearchRoute: string;
   wizardRoute: string;
+  callToActionOptionRoute: string;
   uploadFileRoute: string;
   projectLikes: string;
 }
@@ -39,6 +40,7 @@ export const API_CONFIG: ApiConfig = {
   externalSearchRoute: 'search/external',
   embeddedProjectRoute: 'embed',
   wizardRoute: 'wizard',
+  callToActionOptionRoute: 'callToActionOption',
   uploadFileRoute: 'file',
   projectLikes: 'project/like'
 };