+++ src/app/modules/project/edit/edit.component.ts
@@ -27,6 +27,9 @@ import { AlertConfig } from 'src/app/models/internal/alert-config';
 import { AlertType } from 'src/app/models/internal/alert-type';
 import { AlertService } from 'src/app/services/alert.service';
 import { QuillUtils } from 'src/app/utils/quill.utils';
+import { CallToActionOptionService } from 'src/app/services/call-to-action-option.service';
+import { CallToActionOption } from 'src/app/models/domain/call-to-action-option';
+import { CallToAction } from 'src/app/models/domain/call-to-action';
 import { FileUploaderComponent } from 'src/app/components/file-uploader/file-uploader.component';
 
 /**
@@ -49,8 +52,23 @@ export class EditComponent implements OnInit  {
    * Formgroup for entering project details.
    */
   public editProjectForm: FormGroup;
+  public editCallToActionForm: FormGroup;
   public editCollaboratorForm: FormGroup;
   public project: Project;
+
+  /**
+   * Projects selected call to action
+   */
+  public selectedCallToActionOption: CallToActionOption = {
+    id: 0,
+    type: 'title',
+    value: 'None',
+  };
+
+  /**
+   * The specified redirect url from the call to action
+   */
+  public callToActionRedirectUrl: string;
 
   /**
    * Project's collaborators.
@@ -68,6 +86,13 @@ export class EditComponent implements OnInit  {
   public modulesConfigration = QuillUtils.getDefaultModulesConfiguration();
 
   /**
+   * The available call to action options to select from
+   */
+  public callToActionOptions: CallToActionOption[] = [];
+
+  public callToActionsLoading = true;
+
+  /**
    * Property for storting the invalidId if an invalid project id was entered.
    */
   public invalidId: string;
@@ -82,13 +107,19 @@ export class EditComponent implements OnInit  {
     private formBuilder: FormBuilder,
     private projectService: ProjectService,
     private activatedRoute: ActivatedRoute,
-    private alertService: AlertService
+    private alertService: AlertService,
+    private callToActionOptionService: CallToActionOptionService
   ) {
     this.editProjectForm = this.formBuilder.group({
       name: [null, Validators.required],
       uri: [null, Validators.required],
       shortDescription: [null, Validators.required],
       description: [null],
+    });
+
+    this.editCallToActionForm = this.formBuilder.group({
+      type: [null, Validators.required],
+      value: [null, Validators.required],
     });
 
     this.editCollaboratorForm = this.formBuilder.group({
@@ -108,17 +139,43 @@ export class EditComponent implements OnInit  {
       return;
     }
 
-    this.projectService.get(id)
-      .pipe(
-        finalize(() => this.projectLoading = false)
-      )
-      .subscribe(
-        (result) => {
-          this.project = result;
-          this.collaborators = this.project.collaborators;
-          this.fileUploader.setFiles([this.project.projectIcon]);
-        }
-      );
+    /**
+     * Retrieve the available call to actions
+     */
+    this.callToActionOptionService
+    .getAll()
+    .pipe(finalize(() => (this.callToActionsLoading = false)))
+    .subscribe((result) => {
+        this.callToActionOptions = result.filter(o => o.type === 'title');
+
+        /**
+         * Add the none option to the dropdown
+         */
+        this.callToActionOptions.unshift(this.selectedCallToActionOption);
+
+        this.projectService.get(id)
+          .pipe(
+            finalize(() => this.projectLoading = false)
+          )
+          .subscribe(
+            (projectResult) => {
+              this.project = projectResult;
+              this.collaborators = this.project.collaborators;
+              this.fileUploader.setFiles([this.project.projectIcon]);
+
+              if (this.project.callToAction != null) {
+                for (let i = 0; i < this.callToActionOptions.length; i++) {
+                  const element = this.callToActionOptions[i];
+                  if (element.value.toLowerCase() === this.project.callToAction.optionValue) {
+                    this.selectedCallToActionOption = this.callToActionOptions[i];
+                  }
+                }
+
+                this.callToActionRedirectUrl = this.project.callToAction.value;
+              }
+            }
+          );
+    });
   }
 
   public onClickSubmit(): void {
@@ -136,24 +193,47 @@ export class EditComponent implements OnInit  {
       return;
     }
 
-    const edittedProject: ProjectUpdate = this.editProjectForm.value;
-    edittedProject.collaborators = this.collaborators;
+    const editedProject: ProjectUpdate = this.editProjectForm.value;
+    editedProject.collaborators = this.collaborators;
+
+    /*
+    * Whenever a call to action is selected, this value
+    * should be sent to to the API.
+     */
+    if (this.selectedCallToActionOption.id > 0) {
+      if (this.callToActionRedirectUrl == null) {
+        const alertConfig: AlertConfig = {
+          type: AlertType.danger,
+          preMessage: 'The call to action form is invalid',
+          mainMessage: 'Please add a link to your selected call to action (or set it to "None")',
+          dismissible: true,
+          timeout: this.alertService.defaultTimeout
+        };
+        this.alertService.pushAlert(alertConfig);
+        return;
+      }
+
+      const callToActionToSubmit =
+      { optionValue: this.selectedCallToActionOption.value, value: this.callToActionRedirectUrl } as CallToAction;
+
+      editedProject.callToAction = callToActionToSubmit;
+   }
 
     this.fileUploader.uploadFiles()
       .subscribe(uploadedFiles => {
         if (uploadedFiles) {
           if (uploadedFiles[0]) {
             // Project icon was set and uploaded
-            edittedProject.fileId = uploadedFiles[0].id;
-            this.editProject(edittedProject);
+            editedProject.fileId = uploadedFiles[0].id;
+            this.editProject(editedProject);
           }
           // Project icon was set but not uploaded successfully, the component will show the error
         } else {
           // There was no project icon
           if (this.project.projectIcon) {
-            edittedProject.fileId = 0;
+            editedProject.fileId = 0;
           }
-          this.editProject(edittedProject);
+          this.editProject(editedProject);
         }
       });
   }
@@ -226,7 +306,9 @@ export class EditComponent implements OnInit  {
             type: AlertType.success,
             mainMessage: 'Project was succesfully updated',
             dismissible: true,
-            autoDismiss: true
+            autoDismiss: true,
+            timeout: this.alertService.defaultTimeout
+
           };
           this.alertService.pushAlert(alertConfig);
           this.router.navigate([`project/details/${this.project.id}`]);